Tue, Jul 1, 2025, 2:01 PM -05:00

+++ json_dev/earthquake.h
@@ -7,7 +7,7 @@ typedef struct {
   int timestamp;
   double latitude;
   double longitude;
-  int depth; //km
+  double depth; //km
 } EarthquakeData;
 
 /**
@@ -22,7 +22,7 @@ int initEarthquakeData(EarthquakeData *data,
   int timestamp,
   double latitude,
   double longitude,
-  int depth);
+  double depth);
 
 char *earthquakeDataToString(EarthquakeData *data);
